requirements.txt
flask
flask-cors
google-genai
google-adk
python-dotenv
PyYAML
gunicorn
